# scripts/sdk/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nocodb",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/projects/": {
      "parameters": [],
      "get": {
        "summary": "",
        "operationId": "project-list",
        "responses": {
          "201": {
            "$ref": "#/components/responses/ProjectList"
          }
        },
        "description": "Read project details",
        "requestBody": {
          "$ref": "#/components/requestBodies/ProjectCreate"
        },
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "page"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "pageSize"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "xc-auth",
            "description": "Auth token"
          }
        ],
        "tags": [
          "Meta"
        ]
      },
      "post": {
        "summary": "",
        "operationId": "project-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      }
    },
    "/projects/{projectId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "projectId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "project-read",
        "description": "Read project details",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "xc-auth",
            "description": "Auth token"
          }
        ],
        "tags": [
          "Meta"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "",
        "operationId": "put-projects---copy",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "summary": "",
        "operationId": "delete-projects---copy",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/views/{viewId}/sorts": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "viewId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "sort-list",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SortList"
          },
          "": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      },
      "post": {
        "summary": "",
        "operationId": "sort-create",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sort"
              }
            }
          }
        }
      }
    },
...
    "/views/{viewId}/sorts/{sortId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "viewId",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "sortId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "sort-get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sort"
                }
              }
            }
          }
        },
        "tags": [
          "Meta"
        ]
      },
      "put": {
        "summary": "",
        "operationId": "sort-update",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sort"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "",
        "operationId": "sort-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meta"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectReq": {
        "title": "Project",
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "bases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BaseReq"
            }
          }
        }
      },
      "Project": {
        "title": "Project",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "meta": {
            "type": [
              "string",
              "object"
            ]
          },
          "color": {
            "type": "string"
          },
          "deleted": {
            "type": [
              "string",
              "boolean"
            ]
          },
          "order": {
            "type": "number"
          },
          "bases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Base"
            }
          }
        }
      },
      "ProjectList": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "bases": {
              "list": [
                {
                  "id": "string",
                  "project_id": "string",
                  "alias": "string",
                  "host": "string",
                  "port": 0,
                  "username": "string",
                  "password": "string",
                  "database": "string",
                  "url": "string",
                  "params": "string",
                  "type": "string",
                  "ssl": "string"
                }
              ],
              "pageInfo": {
                "pageSize": 0,
                "totalRows": 0,
                "sort": "string",
                "isFirstPage": true,
                "isLastPage": true
              }
            }
          }
        },
        "properties": {
          "projects": {
            "type": "object",
            "required": [
              "list",
              "pageInfo"
            ],
            "properties": {
              "list": {
                "type": "array",
                "uniqueItems": true,
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/Project"
                }
              },
              "pageInfo": {
                "$ref": "#/components/schemas/Paginated"
              }
            }
          }
        },
        "required": [
          "projects"
        ]
      },
      "Base": {
        "title": "Base",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "database": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "params": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "ssl": {
            "type": "string"
          }
        }
      },
      "BaseReq": {
        "title": "Base",
        "type": "object",
        "properties": {
          "alias": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "database": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "params": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "ssl": {
            "type": "string"
          }
        }
      },
...
      "Sort": {
        "title": "Sort",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fk_model_id": {
            "type": "string"
          },
          "fk_column_id": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        }
      },
      "SortList": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "bases": {
              "list": [
                {
                  "id": "string",
                  "project_id": "string",
                  "alias": "string",
                  "host": "string",
                  "port": 0,
                  "username": "string",
                  "password": "string",
                  "database": "string",
                  "url": "string",
                  "params": "string",
                  "type": "string",
                  "ssl": "string"
                }
              ],
              "pageInfo": {
                "pageSize": 0,
                "totalRows": 0,
                "sort": "string",
                "isFirstPage": true,
                "isLastPage": true
              }
            }
          }
        },
        "properties": {
          "sorts": {
            "type": "object",
            "required": [
              "list"
            ],
            "properties": {
              "list": {
                "type": "array",
                "uniqueItems": true,
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/SharedView"
                }
              }
            }
          }
        },
        "required": [
          "sorts"
        ]
      },
...
      "Paginated": {
        "title": "Paginated",
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer"
          },
          "totalRows": {
            "type": "integer"
          },
          "sort": {
            "type": [
              "string",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/Sort"
            }
          },
          "isFirstPage": {
            "type": "boolean"
          },
          "isLastPage": {
            "type": "boolean"
          },
          "page": {
            "type": "number"
          }
        }
      },
...
      "SharedView": {
        "title": "SharedView",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fk_view_id": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "deleted": {
            "type": "string"
          }
        }
      },
...
    },
    "requestBodies": {
...
      "ProjectCreate": {
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectReq"
                }
              ]
            },
            "examples": {
              "example-1": {
                "value": {
                  "id": "string",
                  "title": "string",
                  "status": "string",
                  "description": "string",
                  "meta": "string",
                  "color": "string",
                  "deleted": "string",
                  "order": 0,
                  "bases": [
                    {
                      "id": "string",
                      "project_id": "string",
                      "alias": "string",
                      "host": "string",
                      "port": 0,
                      "username": "string",
                      "password": "string",
                      "database": "string",
                      "url": "string",
                      "params": "string",
                      "type": "string",
                      "ssl": "string"
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "ProjectList": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProjectList"
            },
            "examples": {
              "example-1": {
                "value": {
                  "projects": "string"
                }
              }
            }
          }
        }
      },
...
      "SortList": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SortList"
            }
          }
        }
      },
...
    }
  }
}